Move update gender identity to seperate file
