+++ VAMobile/src/screens/HomeScreen/ProfileScreen/PersonalInformationScreen/GenderIdentityScreen/GenderIdentityScreen.tsx
@@ -12,7 +12,7 @@ import { logAnalyticsEvent } from 'utils/analytics'
 import { useBeforeNavBackListener, useDestructiveActionSheet, useDowntimeByScreenID, useRouteNavigation, useTheme } from 'utils/hooks'
 import { useDemographics } from 'api/demographics/getDemographics'
 import { useGenderIdentityOptions } from 'api/demographics/getGenderIdentityOptions'
-import { useUpdateGenderIdentity } from 'api/demographics'
+import { useUpdateGenderIdentity } from 'api/demographics/updateGenderIdentity'
 
 type GenderIdentityScreenProps = StackScreenProps<HomeStackParamList, 'GenderIdentity'>
